fixed some boundary case bugs, added input filter

+++ src/Dictionary.java
@@ -1,6 +1,8 @@
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Locale;
+import java.util.regex.*;
 
 /**
  * Dictionary class for dictionary management.
@@ -8,6 +10,9 @@ import java.util.ArrayList;
 public class Dictionary {
   // Since java handles BOM like other chars and db encoding is UTF-8
   public static final String UTF8_BOM = "\uFEFF";
+  //Pattern to filer input data.
+  //Match space(s) at the beginning + ending, characters except a-z, \', <space>, -, and .
+  static Pattern pattern = Pattern.compile("^\s+|[^a-z'. -]|\s+$");
   private static Database database;
   private static final WordTrie trie = new WordTrie();
 
@@ -43,6 +48,14 @@ public class Dictionary {
    * @param vie vie
    */
   public static void addWord(String eng, String vie, String pronounce) throws SQLException {
+    eng = eng.toLowerCase();
+    if (filterInput(eng)) {
+      System.err.println("Invalid word!");
+      return;
+    }
+    if (eng.isEmpty()) {
+      return;
+    }
     trie.insert(eng);
     database.add(eng, vie, " ", pronounce);
   }
@@ -54,10 +67,17 @@ public class Dictionary {
    * @return vietnamese word
    */
   public static String searchWord(String eng) throws SQLException {
+    eng = eng.toLowerCase();
+    if (filterInput(eng)) {
+      return "";
+    }
+    if (eng.isEmpty()) {
+      return "";
+    }
     if (trie.search(eng)) {
       return database.getMeaning(eng);
     }
-    return null;
+    return "";
   }
 
   /**
@@ -66,6 +86,13 @@ public class Dictionary {
    * @param eng eng
    */
   public static void deleteWord(String eng) throws SQLException {
+    eng = eng.toLowerCase();
+    if (filterInput(eng)) {
+      return;
+    }
+    if (eng.isEmpty()) {
+      return;
+    }
     if (trie.search(eng)) {
       trie.delete(trie.root, eng, 0);
       database.delete(eng);
@@ -79,10 +106,14 @@ public class Dictionary {
    * @return array list of String
    */
   public static ArrayList<String> lookUpWord(String eng) {
+    eng = eng.toLowerCase();
+    if (filterInput(eng)) {
+      return new ArrayList<String>();
+    }
     if (eng != null && eng.length() > 0) {
       return trie.getRecommendation(eng);
     }
-    return null;
+    return new ArrayList<String>();
   }
 
   /**
@@ -96,6 +127,11 @@ public class Dictionary {
    */
   public static void changeWord(String oldEng, String oldVie, String newEng, String newVie)
       throws SQLException {
+    newEng = newEng.toLowerCase();
+    if (filterInput(newEng)) {
+      System.err.println("Invalid word!");
+      return;
+    }
     trie.delete(trie.root, oldEng, 0);
     trie.insert(newEng);
     database.delete(oldEng);
@@ -112,7 +148,16 @@ public class Dictionary {
     ResultSet r = database.getAllWords();
     while (r.next()) {
       word = r.getString("word");
+      if (filterInput(word)) {
+        //Deny all words violate the rules.
+        continue;
+      }
       trie.insert(word);
     }
   }
+
+  public static boolean filterInput(String word) {
+    return pattern.matcher(word).find();
+  }
+
 }
